refactor: fix test code smells — dedup MockBackend, strengthen assertions
- Extract shared MockBackend to conftest.py (was duplicated in 3 test files)
- Use isinstance() with runtime_checkable Protocol for conformance tests
  instead of callable(getattr()) which can never fail
- Replace weak assertions with exact values:
  - test_harness: qpos == 0.1 instead of > 0.0, assert step count
  - test_assertions: exact verdict "fail" instead of enum membership
  - test_assertions: exact failure distribution count and constraint names
  - test_visualizer: exact sync_count == 1/2 instead of >= 1/2
