+++ tests/test_runner.py
@@ -5,48 +5,14 @@ from __future__ import annotations
 import json
 import threading
 from pathlib import Path
-from typing import Any
 
 import numpy as np
 
-from roboharness.core.capture import CameraView
 from roboharness.core.harness import Harness, SimulatorBackend
 from roboharness.runner import BatchResult, ParallelTrialRunner, TrialSpec
 from roboharness.storage.task_store import TaskStore, TrialResult
 
-
-class MockBackend:
-    """Thread-safe mock backend for parallel testing."""
-
-    def __init__(self) -> None:
-        self._time = 0.0
-        self._state: dict[str, Any] = {"qpos": [0.0], "qvel": [0.0]}
-
-    def step(self, action: Any) -> dict[str, Any]:
-        self._time += 0.01
-        return self.get_state()
-
-    def get_state(self) -> dict[str, Any]:
-        return {**self._state, "time": self._time}
-
-    def save_state(self) -> dict[str, Any]:
-        return {"state": {**self._state}, "time": self._time}
-
-    def restore_state(self, state: dict[str, Any]) -> None:
-        self._state = {**state["state"]}
-        self._time = state["time"]
-
-    def capture_camera(self, camera_name: str) -> CameraView:
-        return CameraView(name=camera_name, rgb=np.zeros((64, 64, 3), dtype=np.uint8))
-
-    def get_sim_time(self) -> float:
-        return self._time
-
-    def reset(self) -> dict[str, Any]:
-        self._time = 0.0
-        self._state = {"qpos": [0.0], "qvel": [0.0]}
-        return self.get_state()
-
+from .conftest import MockBackend
 
 # -- TrialSpec tests --
